requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-allauth==0.63.3
django-environ==0.11.2
whitenoise==6.6.0
psycopg2-binary==2.9.9
pillow==10.4.0
gunicorn==22.0.0